+++ MusicianFinder.Domain/Models/MusicianLikesStyle.cs
@@ -7,7 +7,7 @@ namespace MusicianFinder.Domain.Models
     public class MusicianLikesStyle
     {
         public long MusicianId { get; private set; }
-        public long StyleId { get; private set; }
+        public int StyleId { get; private set; }
         public bool IsMainStyle { get; private set; }
 
         // Navigation properties
